Fixed bugs and implemented pull all colleges
