+++ ScraperFiles/selenium/Coll of Global Public Health - Grad.json
@@ -1,0 +1,133 @@
+[
+  {
+    "name": "Global Public Health",
+    "yearId": 1204,
+    "classIds": [
+      7411,
+      7413,
+      7414,
+      7415,
+      7416,
+      7417,
+      7418,
+      7427,
+      7544,
+      7545,
+      7446,
+      7447,
+      7462,
+      7463,
+      7460,
+      7470,
+      7471,
+      7472,
+      7473,
+      7474,
+      7475,
+      7476,
+      7477,
+      7458,
+      7461,
+      7481,
+      7482,
+      21129,
+      22911,
+      7419,
+      21131,
+      7459,
+      21128,
+      7449,
+      7529,
+      7536,
+      23619,
+      24003,
+      20675,
+      7497,
+      7489,
+      21124,
+      7445,
+      7441,
+      7457,
+      7516,
+      7517,
+      7479,
+      7522,
+      7444,
+      7469,
+      7511,
+      7537,
+      7442,
+      7431,
+      7523,
+      7524,
+      7525,
+      20665,
+      7432,
+      7433,
+      21240,
+      7542,
+      7420,
+      7421,
+      7422,
+      7437,
+      7468,
+      20676,
+      20685,
+      7439,
+      7440,
+      7483,
+      7484,
+      20674,
+      7423,
+      7438,
+      7450,
+      7488,
+      7538,
+      7436,
+      7515,
+      20678,
+      7501,
+      21197,
+      7448,
+      7502,
+      21133,
+      7434,
+      21020,
+      20981,
+      21857,
+      21858,
+      24714,
+      7412,
+      7443,
+      7507,
+      7508,
+      7451,
+      7429,
+      7466,
+      7491,
+      7485,
+      7486,
+      21334,
+      7453,
+      7454,
+      7455,
+      7456,
+      21337,
+      7452,
+      21938,
+      7514,
+      21130,
+      23804,
+      21021,
+      7430,
+      23975
+    ]
+  },
+  {
+    "name": "Maintenance",
+    "yearId": 1204,
+    "classIds": [
+      7425
+    ]
+  }
+]
