+++ ScraperFiles/selenium/Coll of Global Public Health-UGRD.json
@@ -1,0 +1,34 @@
+[
+  {
+    "name": "Undergrad Global Public Health",
+    "yearId": 1204,
+    "classIds": [
+      17540,
+      17541,
+      17542,
+      17543,
+      17544,
+      20993,
+      17557,
+      17559,
+      17560,
+      17562,
+      17545,
+      17546,
+      17561,
+      17547,
+      17548,
+      17549,
+      21154,
+      17550,
+      17564,
+      17551,
+      17552,
+      17553,
+      17554,
+      17555,
+      17556,
+      17558
+    ]
+  }
+]
